+++ tests/fields/test_html5.py
@@ -314,7 +314,7 @@ class J(Form):
         choices=[
             Choice("https://example.org", "Generic Website"),
             Choice("https://example.org/issues", "Issues Website"),
-            Choice("https://example.org/docs", "Documentation Website"),
+            Choice("https://example.org/docs", "Docs Website"),
         ],
     )
 
@@ -323,21 +323,47 @@ def test_datalist():
     form = J()
     assert (
         form.string5()
-        == """<input id="string5" list="languages" name="string5" type="text" value=""><datalist id="languages"><option value="deu" label="German"><option value="eng" label="English"><option value="fra" label="French"><option value="jpn" label="Japanese"><option value="nld" label="Dutch"><option value="spa" label="Spanish"></datalist>"""
+        == '<input id="string5" list="languages" name="string5" type="text"'
+        ' value="">'
+        '<datalist id="languages">'
+        '<option value="deu" label="German">'
+        '<option value="eng" label="English">'
+        '<option value="fra" label="French">'
+        '<option value="jpn" label="Japanese">'
+        '<option value="nld" label="Dutch">'
+        '<option value="spa" label="Spanish">'
+        '</datalist>'
     )
     assert (
         form.search5()
-        == """<input id="search5" list="languages" name="search5" type="search" value="">"""
+        == '<input id="search5" list="languages" name="search5" type="search"'
+        ' value="">'
     )
     assert (
         form.email5()
-        == """<input id="email5" list="emails" name="email5" type="email" value=""><datalist id="emails"><option value="support@example.org" label="Support Team"><option value="noreply@example.org" label="No Reply"><option value="finance@example.org" label="Financial Department"></datalist>"""
+        == '<input id="email5" list="emails" name="email5" type="email"'
+        ' value="">'
+        '<datalist id="emails">'
+        '<option value="support@example.org" label="Support Team">'
+        '<option value="noreply@example.org" label="No Reply">'
+        '<option value="finance@example.org" label="Financial Department">'
+        '</datalist>'
     )
     assert (
         form.tel5()
-        == """<input id="tel5" list="tels" name="tel5" type="tel" value=""><datalist id="tels"><option value="+316 1234 5678" label="Dutch Office"><option value="+326 1234 5678" label="Belgium Office"><option value="+346 1234 5678" label="Spanish Office"></datalist>"""
+        == '<input id="tel5" list="tels" name="tel5" type="tel" value="">'
+        '<datalist id="tels">'
+        '<option value="+316 1234 5678" label="Dutch Office">'
+        '<option value="+326 1234 5678" label="Belgium Office">'
+        '<option value="+346 1234 5678" label="Spanish Office">'
+        '</datalist>'
     )
     assert (
         form.url5()
-        == """<input id="url5" list="urls" name="url5" type="url" value=""><datalist id="urls"><option value="https://example.org" label="Generic Website"><option value="https://example.org/issues" label="Issues Website"><option value="https://example.org/docs" label="Documentation Website"></datalist>"""
+        == '<input id="url5" list="urls" name="url5" type="url" value="">'
+        '<datalist id="urls">'
+        '<option value="https://example.org" label="Generic Website">'
+        '<option value="https://example.org/issues" label="Issues Website">'
+        '<option value="https://example.org/docs" label="Docs Website">'
+        '</datalist>'
     )